Allow Hotwire.io to be installable as a PWA
PWA = Progressive Web App

Prior to this change, in macos Chrome and Edge, the PWA install prompt
was not visible in the toolbar.

Since both the site.manifest file and the dynamic manifest.json.erb were
listed as the site manifest in the application layout, I believe these
browsers were not consuming the manifest.json.erb file. The
manifest.json.erb file appears to be the desired choice for the site
manifest as it is consistent with the style of web manifest used in
Rails 7.

Removing the site.manifest file allows browsers to pick up the desired
web manifest which contains the necessary properties to make this app
installable as a PWA.

I‘ve also update the names of the relevant icons for the manifest to be
less specific to android chrome.